[refactor] util: split __init__.py into geom and concurrent modules
 Extract 2D geometry functions into a new `geometry.py` module:
 - Rectangle operations: rect_intersect, intersect, normalize_rect, expand_rect,
   is_point_in_rect, rotate_rect, separate_rect_rotation
 - Line/polygon operations: perpendicular_distance, clip_line, get_polygon_bbox,
   slope_of_line, intercept_of_line, project_point_on_line
 - Constants: INSIDE, LEFT, RIGHT, LOWER, UPPER

 Extract concurrency utilities into a new `concurrent.py` module:
 - Rate limiting: limit_invocation
 - Threading: RepeatingTimer, BackgroundWorker
 - Future helpers: executeAsyncTask, bindFuture
 - Decorators: timeout
 - Introspection: inspect_getmembers

General-purpose helpers remain in __init__.py

 All moved symbols are re-exported from `odemis.util` for backward compatibility.
